# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTaskStatusRequest:
    properties:
      color:
        type: string
      is_default:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      order_index:
        minimum: 0
        type: integer
    required:
    - color
    - name
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.TaskStatusDTO:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      order_index:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateTaskStatusRequest:
    properties:
      color:
        type: string
      is_default:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      order_index:
        minimum: 0
        type: integer
    required:
    - color
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API do sistema Organiza Aqui
  title: Organiza Aqui API
  version: "1.0"
paths:
...
  /api/v1/task-statuses:
    get:
      description: Retorna todos os status de tarefas do usuário
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskStatusDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Listar status de tarefas
      tags:
      - task-statuses
    post:
      consumes:
      - application/json
      description: Cria um novo status de tarefa
      parameters:
      - description: Dados do status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskStatusRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TaskStatusDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar status de tarefa
      tags:
      - task-statuses
  /api/v1/task-statuses/{id}:
    delete:
      description: Remove um status de tarefa
      parameters:
      - description: ID do status
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar status de tarefa
      tags:
      - task-statuses
    get:
      description: Retorna um status de tarefa específico
      parameters:
      - description: ID do status
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskStatusDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter status de tarefa
      tags:
      - task-statuses
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um status de tarefa
      parameters:
      - description: ID do status
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskStatusDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar status de tarefa
      tags:
      - task-statuses
  /api/v1/task-statuses/reorder:
    post:
      consumes:
      - application/json
      description: Reordena os status de tarefas do usuário
      parameters:
      - description: Array de IDs de status na ordem desejada
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Reordenar status de tarefas
      tags:
      - task-statuses
...
swagger: "2.0"